+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Version 1.0.1
 
-  * 
+  * Fix Osm::Term date helping methods when term contains nil dates
 
 ## Version 1.0.0